+++ cpp/src/gandiva/CMakeLists.txt
@@ -44,6 +44,7 @@ set(SRC_FILES
     context_helper.cc
     decimal_ir.cc
     decimal_type_util.cc
+    decimal_xlarge.cc
     engine.cc
     date_utils.cc
     expr_decomposer.cc
